# backend/tests/openapi.snapshot.json
{
 "components": {
  "schemas": {
...
   "Body_import_bundle_api_v1_projects_import_bundle_post": {
    "properties": {
     "bundle": {
      "contentMediaType": "application/octet-stream",
      "description": "A PRJ-08 bundle .zip",
      "title": "Bundle",
      "type": "string"
     }
    },
    "required": [
     "bundle"
    ],
    "title": "Body_import_bundle_api_v1_projects_import_bundle_post",
    "type": "object"
   },
...
   "BundleImportResult": {
    "properties": {
     "id_changed": {
      "title": "Id Changed",
      "type": "boolean"
     },
     "needs_relink": {
      "title": "Needs Relink",
      "type": "boolean"
     },
     "project": {
      "$ref": "#/components/schemas/ProjectDetail"
     },
     "roots": {
      "items": {
       "$ref": "#/components/schemas/BundleRoot"
      },
      "title": "Roots",
      "type": "array"
     },
     "source_project_id": {
      "title": "Source Project Id",
      "type": "string"
     }
    },
    "required": [
     "project",
     "source_project_id",
     "id_changed",
     "roots",
     "needs_relink"
    ],
    "title": "BundleImportResult",
    "type": "object"
   },
   "BundleRoot": {
    "description": "One alias of the imported project, checked on this server (PRJ-09).",
    "properties": {
     "root": {
      "$ref": "#/components/schemas/RootInfo"
     },
     "verify": {
      "$ref": "#/components/schemas/VerifyReport"
     }
    },
    "required": [
     "root",
     "verify"
    ],
    "title": "BundleRoot",
    "type": "object"
   },
...
   "FsEntry": {
    "properties": {
     "has_metadata": {
      "default": false,
      "title": "Has Metadata",
      "type": "boolean"
     },
     "kind": {
      "enum": [
       "dir",
       "file"
      ],
      "title": "Kind",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "size": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Size"
     }
    },
    "required": [
     "name",
     "path",
     "kind"
    ],
    "title": "FsEntry",
    "type": "object"
   },
   "FsListing": {
    "properties": {
     "entries": {
      "items": {
       "$ref": "#/components/schemas/FsEntry"
      },
      "title": "Entries",
      "type": "array"
     },
     "parent": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent"
     },
     "path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Path"
     },
     "truncated": {
      "default": false,
      "title": "Truncated",
      "type": "boolean"
     }
    },
    "required": [
     "path",
     "parent",
     "entries"
    ],
    "title": "FsListing",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "LabelEntry": {
    "properties": {
     "color": {
      "pattern": "^#[0-9A-Fa-f]{6}$",
      "title": "Color",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "opacity": {
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Opacity",
      "type": "number"
     },
     "value": {
      "maximum": 65535.0,
      "minimum": 0.0,
      "title": "Value",
      "type": "integer"
     },
     "visible": {
      "default": true,
      "title": "Visible",
      "type": "boolean"
     }
    },
    "required": [
     "value",
     "name",
     "color",
     "opacity"
    ],
    "title": "LabelEntry",
    "type": "object"
   },
...
   "PathRoot": {
    "properties": {
     "alias": {
      "title": "Alias",
      "type": "string"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "role": {
      "default": "source",
      "enum": [
       "source",
       "derived"
      ],
      "title": "Role",
      "type": "string"
     }
    },
    "required": [
     "alias",
     "path"
    ],
    "title": "PathRoot",
    "type": "object"
   },
...
   "ProjectCreate": {
    "properties": {
     "description": {
      "default": "",
      "title": "Description",
      "type": "string"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "preset": {
      "default": "ccrcc",
      "description": "Study preset (PRJ-12)",
      "enum": [
       "ccrcc",
       "generic-ct",
       "none"
      ],
      "title": "Preset",
      "type": "string"
     }
    },
    "required": [
     "name"
    ],
    "title": "ProjectCreate",
    "type": "object"
   },
   "ProjectDetail": {
    "description": "API-03 response: project.json + share link (PRJ-03).",
    "properties": {
     "annotation_sources": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ]
      },
      "title": "Annotation Sources",
      "type": "object"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "default_seg": {
      "default": "imported",
      "title": "Default Seg",
      "type": "string"
     },
     "description": {
      "default": "",
      "title": "Description",
      "type": "string"
     },
     "format": {
      "const": "radiology-workbench-project",
      "default": "radiology-workbench-project",
      "title": "Format",
      "type": "string"
     },
     "format_version": {
      "default": 2,
      "title": "Format Version",
      "type": "integer"
     },
     "label_map": {
      "items": {
       "$ref": "#/components/schemas/LabelEntry"
      },
      "title": "Label Map",
      "type": "array"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "path_roots": {
      "items": {
       "$ref": "#/components/schemas/PathRoot"
      },
      "title": "Path Roots",
      "type": "array"
     },
     "phase_mapping": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Phase Mapping",
      "type": "object"
     },
     "phase_priority": {
      "items": {
       "type": "string"
      },
      "title": "Phase Priority",
      "type": "array"
     },
     "phase_vocabulary": {
      "items": {
       "type": "string"
      },
      "title": "Phase Vocabulary",
      "type": "array"
     },
     "preset": {
      "default": "ccrcc",
      "enum": [
       "ccrcc",
       "generic-ct",
       "none"
      ],
      "title": "Preset",
      "type": "string"
     },
     "project_id": {
      "title": "Project Id",
      "type": "string"
     },
     "segmentations": {
      "items": {
       "$ref": "#/components/schemas/SegmentationSet"
      },
      "title": "Segmentations",
      "type": "array"
     },
     "share_url": {
      "title": "Share Url",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     },
     "viewer_defaults": {
      "$ref": "#/components/schemas/ViewerDefaults"
     }
    },
    "required": [
     "project_id",
     "name",
     "created_at",
     "updated_at",
     "share_url"
    ],
    "title": "ProjectDetail",
    "type": "object"
   },
   "ProjectPatch": {
    "description": "API-03 PATCH body: only provided fields change.",
    "properties": {
     "default_seg": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Seg"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "label_map": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/LabelEntry"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label Map"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "phase_mapping": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "string"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase Mapping"
     },
     "phase_priority": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase Priority"
     },
     "phase_vocabulary": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase Vocabulary"
     },
     "viewer_defaults": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ViewerDefaults"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "title": "ProjectPatch",
    "type": "object"
   },
   "ProjectSummary": {
    "description": "PRJ-02 workspace home row.",
    "properties": {
     "archived": {
      "default": false,
      "title": "Archived",
      "type": "boolean"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "curation_progress": {
      "default": 0.0,
      "title": "Curation Progress",
      "type": "number"
     },
     "last_opened_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Opened At"
     },
     "n_cases": {
      "default": 0,
      "title": "N Cases",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "project_id": {
      "title": "Project Id",
      "type": "string"
     },
     "share_url": {
      "title": "Share Url",
      "type": "string"
     }
    },
    "required": [
     "project_id",
     "name",
     "created_at",
     "share_url"
    ],
    "title": "ProjectSummary",
    "type": "object"
   },
...
   "RelinkResult": {
    "properties": {
     "root": {
      "$ref": "#/components/schemas/RootInfo"
     },
     "verify": {
      "$ref": "#/components/schemas/VerifyReport"
     }
    },
    "required": [
     "root",
     "verify"
    ],
    "title": "RelinkResult",
    "type": "object"
   },
...
   "RootBody": {
    "properties": {
     "path": {
      "minLength": 1,
      "title": "Path",
      "type": "string"
     },
     "role": {
      "anyOf": [
       {
        "enum": [
         "source",
         "derived"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Default: the alias's role, else source",
      "title": "Role"
     }
    },
    "required": [
     "path"
    ],
    "title": "RootBody",
    "type": "object"
   },
   "RootInfo": {
    "description": "API-05 alias row.",
    "properties": {
     "alias": {
      "title": "Alias",
      "type": "string"
     },
     "exists": {
      "title": "Exists",
      "type": "boolean"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "role": {
      "default": "source",
      "enum": [
       "source",
       "derived"
      ],
      "title": "Role",
      "type": "string"
     }
    },
    "required": [
     "alias",
     "path",
     "exists"
    ],
    "title": "RootInfo",
    "type": "object"
   },
...
   "SegmentationSet": {
    "description": "`project.json.segmentations[]` (ADR-0015): one mask per item, per set.",
    "properties": {
     "created_at": {
      "default": "",
      "title": "Created At",
      "type": "string"
     },
     "kind": {
      "default": "imported",
      "enum": [
       "imported",
       "task",
       "manual"
      ],
      "title": "Kind",
      "type": "string"
     },
     "label_mapping": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Label Mapping",
      "type": "object"
     },
     "name": {
      "default": "",
      "title": "Name",
      "type": "string"
     },
     "producer": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SegProducer"
       },
       {
        "type": "null"
       }
      ]
     },
     "seg_id": {
      "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$",
      "title": "Seg Id",
      "type": "string"
     },
     "unmatched": {
      "items": {
       "type": "integer"
      },
      "title": "Unmatched",
      "type": "array"
     }
    },
    "required": [
     "seg_id"
    ],
    "title": "SegmentationSet",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "VerifyReport": {
    "properties": {
     "matched": {
      "default": 0,
      "title": "Matched",
      "type": "integer"
     },
     "mismatched": {
      "default": 0,
      "title": "Mismatched",
      "type": "integer"
     },
     "missing": {
      "default": 0,
      "title": "Missing",
      "type": "integer"
     },
     "sampled": {
      "default": 0,
      "title": "Sampled",
      "type": "integer"
     },
     "samples": {
      "items": {
       "$ref": "#/components/schemas/VerifySample"
      },
      "title": "Samples",
      "type": "array"
     }
    },
    "title": "VerifyReport",
    "type": "object"
   },
...
   "ViewerDefaults": {
    "properties": {
     "layout": {
      "default": "four-up",
      "title": "Layout",
      "type": "string"
     },
     "wl": {
      "default": 50,
      "title": "Wl",
      "type": "number"
     },
     "ww": {
      "default": 400,
      "title": "Ww",
      "type": "number"
     }
    },
    "title": "ViewerDefaults",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Radiology Workbench",
  "version": "3.0.0.dev0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/fs/list": {
   "get": {
    "description": "`role=derived` browses ALLOWED_DERIVED_ROOTS (PRJ-13) instead of ALLOWED_DATA_ROOTS.",
    "operationId": "fs_list_api_v1_fs_list_get",
    "parameters": [
     {
      "in": "query",
      "name": "path",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Path"
      }
     },
     {
      "in": "query",
      "name": "role",
      "required": false,
      "schema": {
       "default": "source",
       "enum": [
        "source",
        "derived"
       ],
       "title": "Role",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FsListing"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Fs List",
    "tags": [
     "fs"
    ]
   }
  },
...
  "/api/v1/projects": {
   "get": {
    "operationId": "list_projects_api_v1_projects_get",
    "parameters": [
     {
      "in": "query",
      "name": "archived",
      "required": false,
      "schema": {
       "default": false,
       "title": "Archived",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ProjectSummary"
         },
         "title": "Response List Projects Api V1 Projects Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Projects",
    "tags": [
     "projects"
    ]
   },
   "post": {
    "operationId": "create_project_api_v1_projects_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Project",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/v1/projects/import-bundle": {
   "post": {
    "description": "API-06 import (PRJ-09): new project from a bundle, plus a per-alias resolve report.",
    "operationId": "import_bundle_api_v1_projects_import_bundle_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_bundle_api_v1_projects_import_bundle_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BundleImportResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Import Bundle",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/v1/projects/{pid}": {
   "get": {
    "operationId": "get_project_api_v1_projects__pid__get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Project",
    "tags": [
     "projects"
    ]
   },
   "patch": {
    "operationId": "patch_project_api_v1_projects__pid__patch",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Project",
    "tags": [
     "projects"
    ]
   }
  },
...
  "/api/v1/projects/{pid}/archive": {
   "post": {
    "operationId": "archive_project_api_v1_projects__pid__archive_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectSummary"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Archive Project",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/v1/projects/{pid}/bundle": {
   "post": {
    "description": "API-06 export: `.zip` of the project folder without `cache/` and never image data.",
    "operationId": "export_bundle_api_v1_projects__pid__bundle_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/zip": {
        "schema": {
         "format": "binary",
         "type": "string"
        }
       }
      },
      "description": "Project bundle (PRJ-08)"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Bundle",
    "tags": [
     "projects"
    ]
   }
  },
...
  "/api/v1/projects/{pid}/roots": {
   "get": {
    "operationId": "list_roots_api_v1_projects__pid__roots_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/RootInfo"
         },
         "title": "Response List Roots Api V1 Projects  Pid  Roots Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Roots",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/v1/projects/{pid}/roots/{alias}": {
   "put": {
    "operationId": "put_root_api_v1_projects__pid__roots__alias__put",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "alias",
      "required": true,
      "schema": {
       "title": "Alias",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RootBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RelinkResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Root",
    "tags": [
     "projects"
    ]
   }
  },
...
  "/api/v1/projects/{pid}/unarchive": {
   "post": {
    "operationId": "unarchive_project_api_v1_projects__pid__unarchive_post",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Unarchive Project",
    "tags": [
     "projects"
    ]
   }
  },
...
 }
}